task_engine_v2/requirements.txt
winrt-runtime>=3.2.1
winrt-Windows.Globalization>=3.2.1
winrt-Windows.Graphics.Imaging>=3.2.1
winrt-Windows.Media.Ocr>=3.2.1
winrt-Windows.Storage>=3.2.1
